[v1.1.0] Migration to pytonlib * Strong refactoring and migrating to pytonlib. * TonlibMultiClient replaced with TonlibManager, old TonlibClient splitted to TonlibClient from pytonlib and TonlibWorker wrapper. * Refactoring of repository structure: not services placed in folders, building and deployment scripts and tools are now in root folder. * Compose files extended to work with Docker Swarm. One of the most powerfull feature is the replication mode for main service. This mode allows service to be scaled horizontally for multiple nodes. * Cache and Logging reworked.
