+++ docker-compose.yaml
@@ -1,24 +1,49 @@
-version: '3.3'
+version: '3.7'
 services:
   main:
+    image: ${DOCKER_REGISTRY:-toncenter}/ton-http-api
     build: 
-      context: .
-      dockerfile: infrastructure/main.Dockerfile
-      args:
-        TON_API_LITE_SERVER_CONFIG: ${TON_API_LITE_SERVER_CONFIG}
-    command: -w ${TON_API_WEBSERVERS_WORKERS} --bind 0.0.0.0:8081
+      context: ton-http-api
+      dockerfile: .docker/Dockerfile
     ports:
-      - ${TON_API_HTTP_PORT}:8081
+      - ${TON_API_HTTP_PORT:-8081}:8081
     environment:
       - TON_API_CACHE_ENABLED
       - TON_API_LOGS_ENABLED
-      - TON_API_LITE_SERVER_CONFIG
+      - TON_API_TONLIB_KEYSTORE
+      - TON_API_TONLIB_LITESERVER_CONFIG
       - TON_API_GET_METHODS_ENABLED
       - TON_API_JSON_RPC_ENABLED
       - TON_API_ROOT_PATH
       - FORWARDED_ALLOW_IPS=${TON_API_FORWARDED_ALLOW_IPS}
-    restart: unless-stopped
+    # restart: unless-stopped
     networks:
       - internal
+    secrets:
+      - liteserver_config
+    command: --host=0.0.0.0 --port=8081 # -w ${TON_API_WEBSERVERS_WORKERS} --bind 0.0.0.0:8081
+    healthcheck:
+      test: curl -sS http://127.0.0.1:8081${TON_API_ROOT_PATH}/healthcheck || echo 1
+      interval: 15s
+      timeout: 3s
+      retries: 12
+    deploy:
+      mode: replicated
+      replicas: 1
+      restart_policy:
+        condition: on-failure
+        delay: 15s
+        max_attempts: 3
+        window: 60s
+      resources:
+        limits:
+          cpus: "1.0"
+          memory: 2G
+        reservations:
+          cpus: "0.25"
+          memory: 256M
+secrets:
+  liteserver_config:
+    file: private/mainnet.json
 networks:
   internal:
