--- index_page/prod/index.html
@@ -1,97 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1"/>
-    <title>TON Center - TON HTTP API</title>
-
-    <style>
-        * {
-            font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
-        }
-
-        html {
-            padding: 0;
-            margin: 0;
-        }
-
-        body {
-            padding: 0;
-            margin-top: 100px;
-            margin-bottom: 50px;
-            margin-left: 20px;
-            margin-right: 20px;
-        }
-
-        .img {
-            width: 90%;
-            max-width: 450px;
-        }
-
-        .center {
-            text-align: center;
-        }
-
-        .h {
-            margin-top: 10px;
-            font-size: 48px;
-            font-weight: bold;
-        }
-
-        .h-desc {
-            margin-top: 10px;
-            font-size: 20px;
-            color: #555;
-        }
-
-        a, a:hover, a:after, a:active, a:focus {
-            color: #1d98dc;
-            text-decoration: none;
-        }
-
-        .body {
-            margin-top: 50px;
-            font-size: 20px;
-            line-height: 150%;
-        }
-
-        .a {
-            margin-top: 50px;
-            font-size: 20px;
-        }
-    </style>
-
-</head>
-<body>
-
-<div class="center">
-    <img class="img" src="coding.svg">
-    <div class="h">TON Center</div>
-    <div class="h-desc">Fast and reliable HTTP API for The Open Network</div>
-
-
-    <div class="body">
-        Using API without API key is limited to 1 request per second.<br>
-        Register your API key in the <a href="https://t.me/tonapibot" target="_blank">@tonapibot</a> to get access
-        with higher limits.
-    </div>
-
-    <div class="a">
-        <a href="/api/v2/">API Documentation</a>
-    </div>
-
-    <div class="a">
-        <a href="https://github.com/toncenter/tonweb" target="_blank">JavaScript SDK</a>
-    </div>
-
-    <div class="a">
-        <a href="https://testnet.toncenter.com/">Testnet API</a>
-    </div>
-
-    <div class="a">
-        <a href="https://github.com/toncenter/ton-http-api" target="_blank">Source Code</a>
-    </div>
-</div>
-
-</body>
-</html>
