--- infrastructure/requirements/main.txt
@@ -1,17 +1,0 @@
-tvm_valuetypes
-aioprocessing
-aioredis
-hiredis
-loguru
-pyyaml
-fastapi
-pydantic
-uvicorn
-pymongo
-crc16
-limits
-pytimeparse
-ring
-redis
-coredis
-gunicorn
