--- infrastructure/requirements/token_bot.txt
@@ -1,3 +1,0 @@
-python-telegram-bot
-pyyaml
-redis
