--- infrastructure/scripts/get_container_logs.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-set -e
-
-LOG_PATH=$(sudo docker inspect --format='{{.LogPath}}' $1)
-mkdir -p logs
-sudo cp $LOG_PATH logs/$1.log
-sudo chmod +r logs/$1.log
