+++ scripts/deploy.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+docker stack deploy -c <(scripts/swarm_config.sh) ton-http-api
