+++ ton-analytics/requirements.txt
@@ -1,0 +1,9 @@
+loguru
+pyyaml
+fastapi
+pydantic
+uvicorn
+pymongo
+pytimeparse
+pandas
+Inject
