+++ ton-http-api/requirements.txt
@@ -1,0 +1,20 @@
+tvm_valuetypes
+aioprocessing
+aioredis
+hiredis
+loguru
+pyyaml
+fastapi
+pydantic
+uvicorn
+pymongo
+crc16
+limits
+requests
+pytimeparse
+ring
+redis
+coredis
+gunicorn
+pytonlib
+inject
